Update singleTransferInstruction.c (Added read and write functions for general register instead of direct register data access

+++ src/registerAndMemory.c
@@ -103,11 +103,11 @@ uint32_t* getMemory() {
     return memory.memory;
 }
 
-uint64_t readGeneral(int regNum, int mode) {
+uint64_t readGeneral(uint64_t regNum, int mode) {
     return generalRegisters.data[regNum];
 };
 
-void writeGeneral(uint32_t regNum, uint64_t data, int mode) {
+void writeGeneral(uint64_t regNum, uint64_t data, int mode) {
     pc.mode = mode;
     if (pc.mode == 32) {
         generalRegisters.data[regNum] = (data >> 32) | 0ULL;
@@ -126,11 +126,6 @@ uint64_t readPC() {
 void writePC32(uint32_t data, int mode) {
     pc.mode = mode;
     pc.data = data;
-//    if (pc.mode == 32) {
-//        pc.data = (data >> 32) | 0ULL;
-//    } else {
-//        pc.data = data;
-//    }
 };
 
 void writePC64(uint64_t data, int mode) {
