+++ src/registerAndMemory.h
@@ -12,8 +12,8 @@
 #include <stdlib.h>
 
 // Function Declaration
-uint64_t readGeneral(int regNum, int mode);
-void writeGeneral(int regNum,uint64_t data, int mode);
+uint64_t readGeneral(uint64_t regNum, int mode);
+void writeGeneral(uint64_t regNum,uint64_t data, int mode);
 uint64_t readPC();
 void writePC32(uint32_t data, int mode);
 void writePC64(uint64_t data, int mode);
